+++ crates/sqlite-wasm-vec/src/lib.rs
@@ -1,0 +1,73 @@
+#![doc = include_str!("../README.md")]
+
+#[allow(non_upper_case_globals)]
+#[allow(non_camel_case_types)]
+#[allow(non_snake_case)]
+mod shim;
+
+#[link(name = "sqlite_vec0")]
+extern "C" {
+    pub fn sqlite3_vec_init();
+}
+
+#[cfg(test)]
+mod tests {
+    use std::ffi::CStr;
+
+    use crate::sqlite3_vec_init;
+    use sqlite_wasm_rs::{
+        sqlite3_auto_extension, sqlite3_close, sqlite3_column_count, sqlite3_column_text,
+        sqlite3_column_type, sqlite3_finalize, sqlite3_open_v2, sqlite3_prepare_v3, sqlite3_step,
+        SQLITE_OK, SQLITE_OPEN_CREATE, SQLITE_OPEN_READWRITE, SQLITE_ROW, SQLITE_TEXT,
+    };
+    use wasm_bindgen_test::wasm_bindgen_test;
+
+    #[wasm_bindgen_test]
+    fn test_auto_extension() {
+        unsafe {
+            sqlite3_auto_extension(Some(std::mem::transmute(sqlite3_vec_init as *const ())));
+        }
+
+        let mut db = std::ptr::null_mut();
+        let ret = unsafe {
+            sqlite3_open_v2(
+                c":memory:".as_ptr().cast(),
+                &mut db as *mut _,
+                SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE,
+                std::ptr::null(),
+            )
+        };
+        assert_eq!(ret, SQLITE_OK);
+
+        let sql = c"select vec_version();";
+        let mut stmt = std::ptr::null_mut();
+        let ret = unsafe {
+            sqlite3_prepare_v3(
+                db,
+                sql.as_ptr().cast(),
+                -1,
+                0,
+                &mut stmt as *mut _,
+                std::ptr::null_mut(),
+            )
+        };
+        assert_eq!(ret, SQLITE_OK);
+
+        unsafe {
+            assert_eq!(sqlite3_step(stmt), SQLITE_ROW);
+            let count = sqlite3_column_count(stmt);
+            assert_eq!(count, 1);
+            let ty = sqlite3_column_type(stmt, 0);
+            assert_eq!(ty, SQLITE_TEXT);
+            let s = CStr::from_ptr(sqlite3_column_text(stmt, 0).cast())
+                .to_str()
+                .unwrap();
+            assert!(s.starts_with('v'));
+            sqlite3_finalize(stmt);
+        }
+
+        unsafe {
+            sqlite3_close(db);
+        }
+    }
+}